#136 - 이미지로더분리 및 비트맵, 비트맵큐브에 1차적용
